refactoring... again
