+++ config.ini
@@ -5,6 +5,7 @@ api_key = for_local_not_needed
 [ai_config]
 persona = lilith_persona.txt
 memory = memory.json
+ai_model = mistral-7b-instruct-v0.3
 temperature = 0.85
 max_tokens = 120
